+++ superagi/apm/analytics_helper.py
@@ -12,23 +12,6 @@ class AnalyticsHelper:
     def __init__(self, session: Session):
         self.session = session
 
-    def create_event(self, event_name: str, event_property: dict, agent_id: int, org_id: int, event_value: int = 1) -> Optional[Event]:
-            try:
-                event = Event(
-                    event_name=event_name,
-                    event_value=event_value,
-                    event_property=event_property,
-                    agent_id=agent_id,
-                    org_id=org_id,
-                )
-                self.session.add(event)
-                self.session.commit()
-                return event
-            except SQLAlchemyError as err:
-                logging.error(f"Error while creating event: {str(err)}")
-                return None
-
-
     def calculate_run_completed_metrics(self) -> Dict[str, Dict[str, Union[int, List[Dict[str, int]]]]]:
 
         agent_model_query = self.session.query(
@@ -211,36 +194,3 @@ class AnalyticsHelper:
         } for row in result]
 
         return running_executions
-
-    def calculate_tool_usage(self) -> List[Dict[str, int]]:
-            tool_usage = []
-            tool_used_subquery = self.session.query(
-                Event.event_property['tool_name'].label('tool_name'),
-                Event.agent_id
-            ).filter_by(event_name="tool_used").subquery()
-
-            agent_count = self.session.query(
-                tool_used_subquery.c.tool_name,
-                func.count(func.distinct(tool_used_subquery.c.agent_id)).label('unique_agents')
-            ).group_by(tool_used_subquery.c.tool_name).subquery()
-
-            total_usage = self.session.query(
-                tool_used_subquery.c.tool_name,
-                func.count(tool_used_subquery.c.tool_name).label('total_usage')
-            ).group_by(tool_used_subquery.c.tool_name).subquery()
-
-            query = self.session.query(
-                agent_count.c.tool_name,
-                agent_count.c.unique_agents,
-                total_usage.c.total_usage
-            ).join(total_usage, total_usage.c.tool_name == agent_count.c.tool_name)
-
-            result = query.all()
-
-            tool_usage = [{
-                'tool_name': row.tool_name,
-                'unique_agents': row.unique_agents,
-                'total_usage': row.total_usage
-            } for row in result]
-
-            return tool_usage
