+++ superagi/controllers/agent.py
@@ -33,7 +33,7 @@ from superagi.models.toolkit import Toolkit
 from sqlalchemy import func
 # from superagi.types.db import AgentOut, AgentIn
 from superagi.helper.auth import check_auth, get_user_organisation
-from superagi.apm.analytics_helper import AnalyticsHelper
+from superagi.apm.event_handler import EventHandler
 
 router = APIRouter()
 
@@ -212,8 +212,8 @@ def create_agent_with_config(agent_with_config: AgentConfigInput,
 
     agent = db.session.query(Agent).filter(Agent.id == db_agent.id,).first()
     organisation = agent.get_agent_organisation(db.session)
-    AnalyticsHelper(session=db.session).create_event('run_created', {'agent_execution_id': execution.id,'agent_execution_name':execution.name}, db_agent.id, organisation.id if organisation else 0),
-    AnalyticsHelper(session=db.session).create_event('agent_created', {'agent_name': agent_with_config.name, 'model': agent_with_config.model}, db_agent.id, organisation.id if organisation else 0)
+    EventHandler(session=db.session).create_event('run_created', {'agent_execution_id': execution.id,'agent_execution_name':execution.name}, db_agent.id, organisation.id if organisation else 0),
+    EventHandler(session=db.session).create_event('agent_created', {'agent_name': agent_with_config.name, 'model': agent_with_config.model}, db_agent.id, organisation.id if organisation else 0)
 
     # execute_agent.delay(execution.id, datetime.now())
